DataCollector/requirements.txt
requests==2.24.0
python-dateutil==2.8.1
python-dotenv==0.14.0
psycopg2==2.8.5